pyproject.toml
[project]
name = "binsketch-algorithm"
version = "0.1.0"
description = "Reproduce binsketch"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "gdown",
    "tqdm",
    "ipykernel",
    "matplotlib",
]

[project.optional-dependencies]
gpu = [
    # For CUDA 11.x, use: cupy-cuda11x
    # For CUDA 12.x, use: cupy-cuda12x
    # Choose based on your NVIDIA CUDA version
    "cupy-cuda12x>=12.0.0",  # Default to CUDA 12.x
]
